app/requirements.txt
celery==5.1.1
Django==3.1.8
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.11.2
redis==3.5.3
django_redis==5.0.0
pyjwt==2.1.0
django-cors-headers==3.7.0
pytz==2020.5
django-currentuser==0.5.3
djoser==2.0.1
django-utils-six==2.0
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
requests==2.26.0